requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
firebase-admin>=6.5.0
google-auth>=2.29.0
python-dotenv>=1.0.1
requests>=2.31.0
httpx>=0.27.0
pydantic>=2.9.0
pydantic-settings>=2.3.0
google-auth-oauthlib>=1.2.0
